refactoring, fix tests

+++ workflow-core/src/main/java/org/opengroup/osdu/workflow/service/WorkflowRunServiceImpl.java
@@ -7,7 +7,9 @@ import java.util.Map;
 import java.util.UUID;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import org.opengroup.osdu.core.common.model.http.AppException;
 import org.opengroup.osdu.core.common.model.http.DpsHeaders;
+import org.opengroup.osdu.workflow.exception.WorkflowRunCompletedException;
 import org.opengroup.osdu.workflow.model.*;
 import org.opengroup.osdu.workflow.provider.interfaces.IWorkflowEngineService;
 import org.opengroup.osdu.workflow.provider.interfaces.IWorkflowMetadataRepository;
@@ -26,6 +28,7 @@ public class WorkflowRunServiceImpl implements IWorkflowRunService {
   private static final String KEY_WORKFLOW_ID = "workflow_id";
   private static final String KEY_CORRELATION_ID = "correlation_id";
   private static final String KEY_EXECUTION_CONTEXT = "execution_context";
+  private static final String KEY_AUTH_TOKEN = "authToken";
   private static final Integer WORKFLOW_RUN_LIMIT = 100;
   private static final ObjectMapper OBJECT_MAPPER = new ObjectMapper();
 
@@ -42,7 +45,7 @@ public class WorkflowRunServiceImpl implements IWorkflowRunService {
   private IWorkflowEngineService workflowEngineService;
 
   @Override
-  public WorkflowRun triggerWorkflow(final String workflowName, final TriggerWorkflowRequest request) {
+  public WorkflowRunResponse triggerWorkflow(final String workflowName, final TriggerWorkflowRequest request) {
     final WorkflowMetadata workflowMetadata = workflowMetadataRepository.getWorkflow(workflowName);
     final String workflowId = workflowMetadata.getWorkflowId();
     final String runId = request.getRunId() != null ? request.getRunId() : UUID.randomUUID().toString();
@@ -50,27 +53,32 @@ public class WorkflowRunServiceImpl implements IWorkflowRunService {
     final WorkflowEngineRequest rq = new WorkflowEngineRequest(runId, workflowId, workflowName);
     final Map<String, Object> context = createWorkflowPayload(workflowId, runId, dpsHeaders.getCorrelationId(), request);
     workflowEngineService.triggerWorkflow(rq, context);
-    return workflowRunRepository.saveWorkflowRun(workflowRun);
+    return buildWorkflowRunResponse(workflowRunRepository.saveWorkflowRun(workflowRun));
   }
 
   @Override
-  public WorkflowRun getWorkflowRunByName(final String workflowName, final String runId) {
+  public WorkflowRunResponse getWorkflowRunByName(final String workflowName, final String runId) {
     WorkflowRun workflowRun = workflowRunRepository.getWorkflowRun(workflowName, runId);
     if (getActiveStatusTypes().contains(workflowRun.getStatus())) {
-      return fetchAndUpdateWorkflowRunStatus(workflowRun);
+      return buildWorkflowRunResponse(fetchAndUpdateWorkflowRunStatus(workflowRun));
     }
-    return workflowRun;
+    return buildWorkflowRunResponse(workflowRun);
   }
 
   @Override
   public void deleteWorkflowRunsByWorkflowName(String workflowName) {
     List<WorkflowRun> workflowRuns = getAllWorkflowRuns(workflowName);
-    List<String> runIdsToDelete = new ArrayList<>();
-    for (WorkflowRun workflowRun : workflowRuns) {
-      runIdsToDelete.add(workflowRun.getRunId());
-    }
-    if (!runIdsToDelete.isEmpty()) {
-      workflowRunRepository.deleteWorkflowRuns(workflowName, runIdsToDelete);
+    if(!isActiveRunsPresent(workflowRuns)) {
+      List<String> runIdsToDelete = new ArrayList<>();
+      for (WorkflowRun workflowRun : workflowRuns) {
+        runIdsToDelete.add(workflowRun.getRunId());
+      }
+      if (!runIdsToDelete.isEmpty()) {
+        workflowRunRepository.deleteWorkflowRuns(workflowName, runIdsToDelete);
+      }
+    } else {
+      String errorMessage = String.format("Active workflow runs found for %s", workflowName);
+      throw new AppException(412, "Failed to delete workflow runs", errorMessage);
     }
   }
 
@@ -81,11 +89,30 @@ public class WorkflowRunServiceImpl implements IWorkflowRunService {
   }
 
   @Override
-  public WorkflowRun updateWorkflowRunStatus(String workflowName, String runId,
+  public WorkflowRunResponse updateWorkflowRunStatus(String workflowName, String runId,
                                              WorkflowStatusType status) {
     WorkflowRun workflowRun = workflowRunRepository.getWorkflowRun(workflowName, runId);
-    workflowRun.setStatus(status);
-    return workflowRunRepository.saveWorkflowRun(workflowRun);
+    if (getCompletedStatusTypes().contains(workflowRun.getStatus())) {
+      throw new WorkflowRunCompletedException(workflowName, runId);
+    } else {
+      if (getActiveStatusTypes().contains(status)) {
+        return buildWorkflowRunResponse(workflowRunRepository.updateWorkflowRun(
+            buildUpdatedWorkflowRun(workflowRun, status, null)));
+      } else {
+        return buildWorkflowRunResponse(workflowRunRepository.updateWorkflowRun(
+            buildUpdatedWorkflowRun(workflowRun, status, System.currentTimeMillis())));
+      }
+    }
+  }
+
+  private boolean isActiveRunsPresent(List<WorkflowRun> workflowRuns) {
+    List<WorkflowStatusType> activeStatusTypes = WorkflowStatusType.getActiveStatusTypes();
+    for(WorkflowRun workflowRun: workflowRuns) {
+      if(activeStatusTypes.contains(workflowRun.getStatus())) {
+        return true;
+      }
+    }
+    return false;
   }
 
   private List<WorkflowRun> getAllWorkflowRuns(String workflowName) {
@@ -109,6 +136,7 @@ public class WorkflowRunServiceImpl implements IWorkflowRunService {
     final Map<String, Object> payload = new HashMap<>();
     payload.put(KEY_RUN_ID, runId);
     payload.put(KEY_WORKFLOW_ID, workflowId);
+    payload.put(KEY_AUTH_TOKEN, dpsHeaders.getAuthorization());
     payload.put(KEY_CORRELATION_ID, correlationId);
     payload.put(KEY_EXECUTION_CONTEXT, OBJECT_MAPPER.convertValue(request.getExecutionContext(), Map.class));
     return payload;
@@ -124,10 +152,10 @@ public class WorkflowRunServiceImpl implements IWorkflowRunService {
         // Setting EndTimeStamp with the timestamp of Instant when this API is called.
         // Currently no EndTimeStamp is returned in the response from Workflow engine.
         // Going forward with the endTimeStamp response from airflow the value can be changed.
-        return workflowRunRepository.updateWorkflowRun(buildWorkflowRunStatus(workflowRun,
+        return workflowRunRepository.updateWorkflowRun(buildUpdatedWorkflowRun(workflowRun,
             currentStatusType, System.currentTimeMillis()));
       } else {
-        return workflowRunRepository.updateWorkflowRun(buildWorkflowRunStatus(workflowRun,
+        return workflowRunRepository.updateWorkflowRun(buildUpdatedWorkflowRun(workflowRun,
             currentStatusType, null));
       }
     }
@@ -146,16 +174,31 @@ public class WorkflowRunServiceImpl implements IWorkflowRunService {
         .build();
   }
 
-  private WorkflowRun buildWorkflowRunStatus(final WorkflowRun workflowRun,
-                                             final WorkflowStatusType workflowStatusType,
-                                             final Long workflowRunEndTimeStamp) {
+  private WorkflowRunResponse buildWorkflowRunResponse(final WorkflowRun workflowRun) {
+    if(workflowRun == null) {
+      return null;
+    }
+    return WorkflowRunResponse.builder()
+        .workflowId(workflowRun.getWorkflowId())
+        .runId(workflowRun.getRunId())
+        .startTimeStamp(workflowRun.getStartTimeStamp())
+        .endTimeStamp(workflowRun.getEndTimeStamp())
+        .submittedBy(workflowRun.getSubmittedBy())
+        .status(workflowRun.getStatus())
+        .build();
+  }
+
+  private WorkflowRun buildUpdatedWorkflowRun(final WorkflowRun workflowRun,
+      final WorkflowStatusType workflowStatusType,
+      final Long workflowRunEndTimeStamp) {
     return WorkflowRun.builder()
+        .workflowId(workflowRun.getWorkflowId())
         .runId(workflowRun.getRunId())
         .startTimeStamp(workflowRun.getStartTimeStamp())
         .endTimeStamp(workflowRunEndTimeStamp)
         .submittedBy(workflowRun.getSubmittedBy())
+        .workflowEngineExecutionDate(workflowRun.getWorkflowEngineExecutionDate())
         .status(workflowStatusType)
-        .workflowId(workflowRun.getWorkflowId())
         .workflowName(workflowRun.getWorkflowName())
         .build();
   }
